again

+++ tobbo.kirill/I2/lab2.cpp
@@ -2,6 +2,7 @@
 #include <fstream>
 #include <string>
 #include <sstream>
+#include <cmath>
 #include "methods.h"
 
 void process_static_array();
@@ -55,7 +56,7 @@ void process_array(std::string prefix, int* arr, size_t size)
   std::cout << "Max value=" << maximum << " count of max: " << count_maximum << "\n";
 
   print_array("before: ", arr, size);
-  shift_part_to_end(arr, size, 0, std::ceil(size / 2));
+  shift_part_to_end(arr, size, 0, static_cast<size_t>(std::ceil(size / 2)));
   print_array("after:  ", arr, size);
 
   std::cout << "-----------------\n";
